% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasPlot2d.R
\name{AtlasPlot2d}
\alias{AtlasPlot2d}
\title{2D atlas plot}
\usage{
AtlasPlot2d(data, atlas, fill, groupby = NULL, ...)
}
\arguments{
\item{data}{A dataframe containing variables to be plotted. Must contain a
"label" columns matching the corresponding column in a ggseg atlas file.}

\item{atlas}{A ggseg atlas file.}

\item{fill}{String denoting the column name in data to be plotted.}

\item{groupby}{If facetting, variable to facet by.}

\item{...}{Additional arguments to be passed to ggseg}
}
\value{
A ggseg plot object
}
\description{
Wrapper around ggseg which handles some necessary data manipulation.
}
