% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasPlot3d.R
\name{AtlasPlot3d}
\alias{AtlasPlot3d}
\title{3D surface plot}
\usage{
AtlasPlot3d(
  data,
  atlas,
  surf = "inflated",
  hemi = "right",
  fill,
  palette = "r_RdBu",
  ncolor = 9,
  limits = NULL,
  camera = "right lateral",
  ...
)
}
\arguments{
\item{data}{A dataframe containing variables to be plotted. Must contain a
"label" columns matching the corresponding column in a ggseg atlas file.}

\item{atlas}{A ggseg 3D atlas file. Currently only supports schaefer17_400_3d
from ggsegSchaefer.}

\item{surf}{Plotting surface. Defaults top "inflated".}

\item{hemi}{Cortical hemisphere.}

\item{fill}{String denoting the column name in data to be plotted.}

\item{palette}{Name of RColorBrewer palette. Prepend with "r_" to reverse
color scale. Default is set to r_RdBu.}

\item{limits}{Colorscale limits. If only one (positive) value c is supplied,
colorscale is assumed to be symmetric with limits (-c,c). If NULL, scale is set
to (-c,c) where c = max(abs(fill)).}

\item{camera}{Camera angle. Defaults to "right lateral".}

\item{...}{Additional arguments to be passed to ggseg3d}
}
\value{
A ggseg3d plot object
}
\description{
Wrapper around ggseg3d which handles some necessary data manipulation, and
includes more convenient handling of colormaps.
}
