% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CortexPlot2d.R
\name{CortexPlot2d}
\alias{CortexPlot2d}
\title{2D surface plot}
\usage{
CortexPlot2d(data, atlas, fill, groupby = NULL, ...)
}
\arguments{
\item{data}{A dataframe containing variables to be plotted. Must contain a
"label" columns matching the corresponding column in a ggseg atlas file.}

\item{atlas}{A ggseg 3D atlas file. Currently only supports schaefer17_400_3d
from ggsegSchaefer.}

\item{fill}{String denoting the column name in data to be plotted.}

\item{groupby}{If facetting, variable to facet by.}

\item{...}{Additional arguments to be passed to ggseg}
}
\value{
A ggseg plot object
}
\description{
Wrapper around ggseg3d which handles some necessary data manipulation, and
includes more convenient handling of colormaps.
}
