% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAtlasInformation.R
\name{LoadAtlasInformation}
\alias{LoadAtlasInformation}
\title{Load atlas file}
\usage{
LoadAtlasInformation(atlas)
}
\arguments{
\item{atlas}{Name of atlas (see details)}
}
\value{
A data frame
}
\description{
Load ROI information associated with an atlas.
}
\details{
Current supports the following atlases:
\itemize{
 \item{"schaefer17_400"}{Schaefer 400 region parcelation (Yeo networks).}
 \item{"SUIT28cb"}{28 regions cerebellar parcellation modelled after the
 SUIT atlas described by Diedrichsen (2006).}
 \item{"Subcort14bg"}{Rudimentary parcellation of the basal ganglia.}
}
}
