% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Subcort14bg}
\alias{Subcort14bg}
\title{Subcort14bg}
\format{
An S3 object of class brain_atlas
}
\usage{
Subcort14bg
}
\description{
Rudimentary atlas of the basal ganglia.
}
\keyword{datasets}
