% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SUIT28cb}
\alias{SUIT28cb}
\title{SUIT28cb}
\format{
An S3 object of class brain_atlas
}
\usage{
SUIT28cb
}
\description{
Cerebelar coordinate data modelled after the SUIT atlas described in
Diedrichsen (2006).
}
\references{
Diedrichsen, J. (2006) Neuroimage, 33, 1, p. 127-138.
}
\keyword{datasets}
