% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAtlasInformation.R
\name{LoadAtlasInformation}
\alias{LoadAtlasInformation}
\title{Load atlas file}
\usage{
LoadAtlasInformation(atlas)
}
\arguments{
\item{atlas}{Name of atlas (currently only supports schaefer17_400)}
}
\value{
A data frame
}
\description{
Load ROI information associated with an atlas. Currently supports only the
Schaefer 400 parcel 17 Yeo-network parcellation.
}
